binance_robot/requirements.txt
#pip install -r requirements.txt
python-binance
pandas
time
